langchain/requirements.txt
langchain-openai
langchain-core
langchain
python-dotenv